Got the basic e10s stuff working (sort of)
Functionally, this is only missing the cleanup part when the add-on is
removed. Since the global message manager can't send messages, there
needs to be a bit of an interation on all opened browsers to get this
done. The message listener is already in place, though.

This code works as expected with e10s turned off. When it is on, it's
only working after a browser restart (this is on Mac OS X). On
installation, I'm getting a couple of these errors, which might be
related:

aDocShell is null browser.js:157

However, even on new tabs this isn't working, so this looks like a
Firefox bug.